@@ -50,7 +50,8 @@ namespace ResearchXBRL.Infrastructure.Services.EdinetXBRLParser
                 .Where(x => x.GetAttributeValue("unitRef") is not null)
                 .Where(x =>
                     x.Name.StartsWith("jppfs_cor:")
-                 || x.Name.StartsWith("jpigp_cor:"))
+                 || x.Name.StartsWith("jpigp_cor:")
+                 || x.Name.StartsWith("jpcrp_cor:"))
                 .Select(CreateReportItem);
         }
 
